feat: Added Authentication to application

+++ AirBNB/src/main/java/com/example/AirBNB/service/RoomServiceImpl.java
@@ -4,12 +4,17 @@ package com.example.AirBNB.service;
 import com.example.AirBNB.dto.RoomDto;
 import com.example.AirBNB.entity.Hotel;
 import com.example.AirBNB.entity.Room;
+import com.example.AirBNB.entity.User;
 import com.example.AirBNB.exception.ResourceNotFoundException;
+import com.example.AirBNB.exception.UnAuthorisedException;
 import com.example.AirBNB.repository.HotelRepository;
 import com.example.AirBNB.repository.RoomRepository;
+import com.example.AirBNB.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,14 +30,21 @@ public class RoomServiceImpl implements RoomService{
     private final HotelRepository hotelRepository;
     private final InventoryService inventoryService;
     private final ModelMapper modelMapper;
+    private final UserRepository userRepository;
 
     @Override
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
         log.info("Creating a new room in hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository
                 .findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
         Room room = modelMapper.map(roomDto, Room.class);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         room.setHotel(hotel);
         room = roomRepository.save(room);
 
@@ -49,6 +61,12 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+ hotelId);
+        }
+
 
         return hotel.getRooms()
                 .stream()
@@ -72,6 +90,13 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+ room.getHotel().getId());
+        }
+
+
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
     }
